+++ apps/console/src/features/authentication-flow-builder/components/script-editor-panel/__tests__/__mocks__/permissions.ts
