+++ apps/console/src/features/authentication-flow-builder/components/script-editor-panel/__tests__/script-editor-side-panel.test.tsx
