+++ apps/console/src/features/authentication-flow-builder/components/script-editor-panel/__tests__/secret-delete-confirmation-modal.test.tsx
