+++ apps/console/src/features/authentication-flow-builder/components/script-editor-panel/adaptive-script-reset-confirmation-modal.tsx
