+++ apps/console/src/features/authentication-flow-builder/components/script-editor-panel/script-based-flow-switch.scss
