+++ apps/console/src/features/authentication-flow-builder/components/script-editor-panel/script-editor-panel.scss
@@ -16,12 +16,8 @@
  * under the License.
  */
 
-.script-editor-side-panel {
+.script-editor-panel {
     height: 100%;
-
-    .predefined-flows-side-panel-content {
-        border-left: 1px solid var(--oxygen-palette-divider);
-    }
 
     .toolbar-container {
         .oxygen-toolbar {
