+++ apps/console/src/features/authentication-flow-builder/components/script-editor-panel/secret-delete-confirmation-modal.tsx
