+++ apps/console/src/features/authentication-flow-builder/components/script-editor-panel/secret-selection-dropdown.scss
